Add files via upload
This is the first version of the texting grocery list. There's a few things to note:
You'll want to have a Twilio account set up with a phone number(it's free)
You'll want to have a csv file like the one above. 
To delete a menu item, you have to manually delete it by opening up the csv file and deleting the line


I'd like to work on this to make it so you can delete lists and enter your number through the program instead of in the code manually.